Backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
yfinance
numpy
pandas
scikit-learn
python-dotenv